# perf: reduces runtime prototype traversal for hasOwnProperty