@enisdenjo 2020-11-04
:tada: This PR is included in version 1.11.0 :tada:

The release is available on:
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/graphql-ws/v/1.11.0)
- [GitHub release](https://github.com/enisdenjo/graphql-ws/releases/tag/v1.11.0)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: